# Refactor featured queries