--- a/templates/articles/article_landing_page.html
+++ b/templates/articles/article_landing_page.html
@@ -19,54 +19,60 @@
       </div>
       <div class="row">
         <div class="col-12">
-          {% for item in self.featured_xlarge %}
-            {% include "includes/features/feature_content_xlarge.html" with content=item.article_page.specific %}
-          {% endfor %}
-        </div>
-      </div>
-      <hr>
-      <div class="row">
-        <div class="col-12 col-md-8">
-          {% for item in self.featured_large_1 %}
-            {% include "includes/features/feature_content_large.html" with content=item.article_page.specific %}
-          {% endfor %}
-        </div>
-        <hr>
-        <div class="col-12 col-md-4 featured-small">
-          {% for item in self.featured_small_1 %}
-            {% if forloop.counter > 1 %}
-              <hr>
-            {% endif %}
-            {% include "includes/features/feature_content_small.html" with content=item.article_page.specific %}
-          {% endfor %}
-        </div>
-      </div>
-      <hr>
-      <div class="row">
-        {% for item in self.featured_medium_1 %}
-          {% if forloop.counter > 1 %}
-            <hr>
+          {% if featured_articles|length > 0 %}
+            {% include "includes/features/feature_content_xlarge.html" with content=featured_articles.0 %}
           {% endif %}
-          <div class="col-12 col-md-6">
-            {% include "includes/features/feature_content_medium.html" with content=item.article_page.specific with_subtitle=True %}
-          </div>
-        {% endfor %}
+        </div>
+      </div>
+      <hr>
+      <div class="row">
+        <div class="col-12 col-md-8">
+          {% if featured_articles|length > 1 %}
+            {% include "includes/features/feature_content_large.html" with content=featured_articles.1 %}
+          {% endif %}
+        </div>
+        <hr>
+        <div class="col-12 col-md-4 featured-small">
+          {% with featured_articles|slice:"2:7" as featured_small %}
+            {% for item in featured_small %}
+              {% if forloop.counter > 1 %}
+                <hr>
+              {% endif %}
+              {% include "includes/features/feature_content_small.html" with content=item %}
+            {% endfor %}
+          {% endwith %}
+        </div>
+      </div>
+      <hr>
+      <div class="row">
+        {% with featured_articles|slice:"7:9" as featured_medium %}
+          {% for item in featured_medium %}
+            {% if forloop.counter > 1 %}
+              <hr>
+            {% endif %}
+            <div class="col-12 col-md-6">
+              {% include "includes/features/feature_content_medium.html" with content=item with_subtitle=True %}
+            </div>
+          {% endfor %}
+        {% endwith %}
       </div>
       <hr>
       <div class="row">
         <div class="col-12 col-md-4 featured-small">
-          {% for item in self.featured_small_2 %}
-            {% if forloop.counter > 1 %}
-              <hr>
-            {% endif %}
-            {% include "includes/features/feature_content_small.html" with content=item.article_page.specific %}
-          {% endfor %}
+          {% with featured_articles|slice:"10:15" as featured_small %}
+            {% for item in featured_small %}
+              {% if forloop.counter > 1 %}
+                <hr>
+              {% endif %}
+              {% include "includes/features/feature_content_small.html" with content=item %}
+            {% endfor %}
+          {% endwith %}
         </div>
         <hr>
         <div class="col-12 col-md-8">
-          {% for item in self.featured_large_2 %}
-            {% include "includes/features/feature_content_large.html" with content=item.article_page.specific %}
-          {% endfor %}
+          {% if featured_articles|length > 9 %}
+            {% include "includes/features/feature_content_large.html" with content=featured_articles.9 %}
+          {% endif %}
         </div>
       </div>
     </div>