--- a/templates/articles/article_landing_page.html
+++ b/templates/articles/article_landing_page.html
@@ -83,7 +83,7 @@
         <div class="row">
           <div class="col-12">
             <h2 class="paragraph-heading">Series</h2>
-            {% include "includes/swiper.html" with swiper_content=self.all_article_series type="article-landing-series" %}
+            {% include "includes/swiper.html" with swiper_content=featured_article_series type="article-landing-series" %}
           </div>
         </div>
       </div>