--- a/templates/home/home_page.html
+++ b/templates/home/home_page.html
@@ -73,7 +73,7 @@
   <section class="homepage-promotion-blocks">
     <div class="container">
       <div class="row">
-        {% for promotion_block in self.promotion_blocks_list %}
+        {% for promotion_block in promotion_blocks %}
           <div class="homepage-promotion-block col-md-6">
             {% include "includes/promotion_block.html" with promotion_block=promotion_block %}
           </div>