Refactored the featured queries on the Opinions, Multimedia, and Events pages to match the optimizations done to the home page in #1170.

See the before/after table for response time and queries. Note that the home page was only optimized to update the promotion blocks query, which resulted in removing just a single query.

Page | Before - Time | Before - Queries | After - Time | After - Queries
--- | --- | --- | --- | ---
Home | 10714.93ms | 175 queries in 265.54ms | 10397.40ms | 174 queries in 271.81ms
Opinions | 9065.69ms | 183 queries in 217.63ms | 5820.57ms | 111 queries in 116.16ms
Multimedia | 2664.04ms | 54 queries in 58.63ms | 1853.61ms | 34 queries in 51.02ms
Events | 4101.96ms | 31 queries in 178.17ms | 1463.13ms | 22 queries in 39.37ms